requirements.txt
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
python-dotenv==1.0.0
numpy==1.26.2
plotly==5.24.1
anthropic>=0.40.0
